requirements.txt
python-dotenv==1.0.0
spotipy==2.24.0
ytmusicapi==1.9.0
pysocks==1.7.1